everything has 100% except 2 files, (above 80%) todo
