--- docs/notes.md
@@ -1,4 +1,0 @@
-[Heuristic Methods for Large Centroid Clustering Problems](https://link.springer.com/article/10.1023/A:1021841728075)
-
-private npm repo that seems to support scoped packages like @types/
-https://github.com/verdaccio/verdaccio
