+++ jest.config.js
@@ -1,6 +1,11 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 module.exports = {
     collectCoverage: true,
+    collectCoverageFrom: ['src/**/*'],
+    coveragePathIgnorePatterns: ['node_modules', 'test', 'doc.ts'],
+    coverageDirectory: 'coverage',
+    coverageProvider: 'v8',
+    coverageReporters: ['json', 'lcov', 'text', 'clover'],
     preset: 'ts-jest',
     testEnvironment: 'node',
     verbose: true,
