+++ test/level1/complex/level1.complex.test.ts
@@ -17,11 +17,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 //import { expect } from 'chai';
 import * as blas from '../../../src/lib';
-import { approximately } from '../../test-helpers';
+import { approximately, each, multiplexer, fortranArrComplex64, muxCmplx } from '../../test-helpers';
 import { fixture } from './fixtures';
 
 const {
-    util: { each, multiplexer, fortranArrComplex64, muxCmplx },
     level1: { scasum, cswap, csscal, caxpy, ccopy, cdotc, cdotu, crotg, cscal, csrot, icamax },
 } = blas;
