+++ test/level3/complex/csymm/test.ts
@@ -15,13 +15,11 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import * as blas from '../../../../src/lib';
-import { approximatelyWithPrec } from '../../../test-helpers';
-
 import { fixture } from './fixtures';
+import * as blas from '../../../../src/lib';
+import { approximatelyWithPrec, each, multiplexer, fortranMatrixComplex64 } from '../../../test-helpers';
 
 const {
-    util: { each, multiplexer, fortranMatrixComplex64 },
     level3: { csymm },
 } = blas;
