+++ test/level3/double/ssyr2k/test.ts
@@ -15,13 +15,11 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-import * as blas from '../../../../src/lib';
-import { real } from '../../../../src/lib/f_func';
-import { approximatelyWithPrec } from '../../../test-helpers';
 import { fixture } from './fixtures';
+import * as blas from '../../../../src/lib';
+import { approximatelyWithPrec, real, each, multiplexer, fortranMatrixComplex64 } from '../../../test-helpers';
 
 const {
-    util: { each, multiplexer, fortranMatrixComplex64 },
     level3: { ssyr2k },
 } = blas;
