+++ tsconfig.json
@@ -81,7 +81,9 @@
         // we dont use dynamic module loading or the ES6 module spec
         "moduleResolution": "node",
         "paths": { //  we can use this for the test to pick up the module
-            "blasjs": [ "src/lib" ]
+            "@blasjs/*": [
+                "src/lib/*"
+            ]
         }, // maybe use later, if you decide to use this, specify baseurl
         // source maps
         "preserveSymlinks": true, // keep it around just in case
